Renamed module sirius -> authentication
