+++ lambdas/data_providers/authentication/__init__.py
@@ -1,4 +1,4 @@
 class SiriusAuthenticationError(Exception):
     pass
 
-from .authentication import SiriusAuthenticator
+from .sirius import SiriusAuthenticator
