+++ lambdas/tests/data_providers/authentication/test_sirius.py
@@ -4,7 +4,7 @@ import pytest
 import responses
 from unittest import mock
 from requests import Request
-from data_providers.sirius import SiriusAuthenticator, SiriusAuthenticationError
+from data_providers.authentication import SiriusAuthenticator, SiriusAuthenticationError
 
 
 class TestSiriusAuthenticator(object):
@@ -33,7 +33,7 @@ class TestSiriusAuthenticator(object):
     # -------------------------------------
     # Object tests
 
-    @mock.patch('data_providers.sirius.authentication.boto3', autospec=True)
+    @mock.patch('data_providers.authentication.sirius.boto3', autospec=True)
     @mock.patch.dict('os.environ', {'URL_MEMBRANE': 'https://example.com'})
     @mock.patch.dict('os.environ', {'DYNAMODB_AUTH_CACHE_TABE_NAME': 'tabme-name'})
     @mock.patch.dict('os.environ', {'CREDENTIALS': json.dumps({'email': 'test@example.com', 'password': 'password'})})
@@ -71,7 +71,7 @@ class TestSiriusAuthenticator(object):
                 Item={'id': 'token','token': test_token}
             )
 
-    @mock.patch('data_providers.sirius.authentication.boto3', autospec=True)
+    @mock.patch('data_providers.authentication.sirius.boto3', autospec=True)
     @mock.patch.dict('os.environ', {'URL_MEMBRANE': 'https://example.com'})
     @mock.patch.dict('os.environ', {'DYNAMODB_AUTH_CACHE_TABE_NAME': 'tabme-name'})
     @mock.patch.dict('os.environ', {'CREDENTIALS': json.dumps({'email': 'test@example.com', 'password': 'password'})})
@@ -98,7 +98,7 @@ class TestSiriusAuthenticator(object):
             mock_boto3.resource.return_value.Table.return_value.get_item.assert_not_called()
             mock_boto3.resource.return_value.Table.return_value.put_item.assert_not_called()
 
-    @mock.patch('data_providers.sirius.authentication.boto3', autospec=True)
+    @mock.patch('data_providers.authentication.sirius.boto3', autospec=True)
     @mock.patch.dict('os.environ', {'URL_MEMBRANE': 'https://example.com'})
     @mock.patch.dict('os.environ', {'DYNAMODB_AUTH_CACHE_TABE_NAME': 'tabme-name'})
     @mock.patch.dict('os.environ', {'CREDENTIALS': json.dumps({'email': 'test@example.com', 'password': 'password'})})
@@ -129,7 +129,7 @@ class TestSiriusAuthenticator(object):
             # But as that fails, we don't try and add anything to the cache.
             mock_boto3.resource.return_value.Table.return_value.put_item.assert_not_called()
 
-    @mock.patch('data_providers.sirius.authentication.boto3', autospec=True)
+    @mock.patch('data_providers.authentication.sirius.boto3', autospec=True)
     @mock.patch.dict('os.environ', {'URL_MEMBRANE': 'https://example.com'})
     @mock.patch.dict('os.environ', {'DYNAMODB_AUTH_CACHE_TABE_NAME': 'tabme-name'})
     @mock.patch.dict('os.environ', {'CREDENTIALS': json.dumps({'email': 'test@example.com', 'password': 'password'})})
@@ -169,7 +169,7 @@ class TestSiriusAuthenticator(object):
             assert 'HTTP-SECURE-TOKEN' in r.headers
             assert r.headers['HTTP-SECURE-TOKEN'] == test_token
 
-    @mock.patch('data_providers.sirius.authentication.boto3', autospec=True)
+    @mock.patch('data_providers.authentication.sirius.boto3', autospec=True)
     @mock.patch.dict('os.environ', {'URL_MEMBRANE': 'https://example.com'})
     @mock.patch.dict('os.environ', {'DYNAMODB_AUTH_CACHE_TABE_NAME': 'tabme-name'})
     @mock.patch.dict('os.environ', {'CREDENTIALS': json.dumps({'email': 'test@example.com', 'password': 'password'})})
